Add more tests and ci steps

+++ src/client/src/components/UserForm/index.tsx
@@ -1,17 +1,16 @@
 import React, { Component, SyntheticEvent } from 'react';
-import { Formik, FormikActions } from "formik";
+import { Formik, FormikActions, FormikConfig, FormikProps } from 'formik';
 import Form from './Form';
 import {userForm} from '../../validation';
 import { routes } from '../../config';
-import { History } from 'history';
 import { IUsersState, ICreatedUser, IUser } from '../../reducers/initialState';
+import { ConnectedRouterProps } from 'connected-react-router';
 
-export interface IProps {
+export interface IProps extends ConnectedRouterProps {
   onEntitySave: (payload: IUser|ICreatedUser) => void;
   onDelete?: (id: string) => void;
-  history: History;
   users: IUsersState;
-  id: string;
+  id?: string;
 }
 
 class UserForm extends Component<IProps> {
@@ -22,7 +21,6 @@ class UserForm extends Component<IProps> {
 
       this.props.history.push(routes.users.list);
     } catch (error) {
-      console.log('error', error);
       actions.setSubmitting(false);
       actions.setErrors(error);
       actions.setStatus({ msg: 'Set some arbitrary status or data' });
@@ -31,14 +29,25 @@ class UserForm extends Component<IProps> {
 
   handleDelete = (event: SyntheticEvent) => {
     event.preventDefault();
-    if (this.props.onDelete) {
+    if (this.props.onDelete && this.props.id) {
       this.props.onDelete(this.props.id);
     }
     this.props.history.push(routes.users.list);
   }
 
+  handleGoBack = () => this.props.history.goBack();
+
+  handleFormikRender = (props: FormikProps<IUser>) => (
+    <Form
+      {...props}
+      onDelete={this.handleDelete}
+      onGoBack={this.handleGoBack}
+      id={this.props.id}
+    />
+  )
+
   render() {
-    const item = this.props.users.items[this.props.id] || {};
+    const item = this.props.users.items[`${this.props.id}`] || {};
 
     const values: IUser|ICreatedUser = {
       id: item.id,
@@ -61,12 +70,7 @@ class UserForm extends Component<IProps> {
 
     return (
       <Formik
-        render={props => <Form
-          {...props}
-          history={this.props.history}
-          onDelete={this.handleDelete}
-          id={this.props.id}
-        />}
+        render={this.handleFormikRender}
         initialValues={values}
         validationSchema={userForm}
         onSubmit={this.handleSubmit}
